+++ docs/CHANGELOG.md
@@ -39,3 +39,4 @@
 
 - Added custom validation to <fc-select>, removed strict validation from it.
 - Split toggledropdown function on <fc-select> into hideDropdown and showDropdown
+- Updated <fc-select> CSS Tokens.